fix: sidebar & access control

+++ admin/eventUpload.php
@@ -1,39 +1,44 @@
 <?php
-require('../config.php');
-session_start();
+    require('../config.php');
+    session_start();
 
-$id = $_SESSION['id'];
+    $id = $_SESSION['id'];
 
-if ($_SERVER['REQUEST_METHOD'] == 'POST' && isset($_FILES['eventBanner'])) {
-    $eventBanner = $_FILES['eventBanner']['name'];
-    $tmp = $_FILES['eventBanner']['tmp_name'];
+    if (!isset($_SESSION['id']) || $_SESSION['role'] !== 'admin') {
+        header('Location: ../auth/login.php');
+        exit();
+    }
 
-    // Fetch current profile picture from the database
-    $stmt = $conn->prepare("SELECT event_banner FROM event WHERE event_id = ?");
-    $stmt->execute([$event_id]);
-    $currentPhoto = $stmt->fetchColumn();
+    if ($_SERVER['REQUEST_METHOD'] == 'POST' && isset($_FILES['eventBanner'])) {
+        $eventBanner = $_FILES['eventBanner']['name'];
+        $tmp = $_FILES['eventBanner']['tmp_name'];
 
-    // Check for errors
-    if ($_FILES['eventBanner']['error'] == 0) {
-        // Delete the old photo if it exists
-        if ($currentPhoto && file_exists("../uploads/eventBanner/$currentPhoto")) {
-            unlink("../uploads/eventBanner/$currentPhoto");
-        }
-        
-        if (move_uploaded_file($tmp, '../uploads/eventBanner/' . $eventBanner)) {
-            $query = "UPDATE event SET event_banner = ? WHERE event_id = ?";
-            $stmt = $conn->prepare($query);
-            $stmt->execute([$eventBanner, $event_id]);
+        // Fetch current profile picture from the database
+        $stmt = $conn->prepare("SELECT event_banner FROM event WHERE event_id = ?");
+        $stmt->execute([$event_id]);
+        $currentPhoto = $stmt->fetchColumn();
 
-            $_SESSION['success'] = "Profile picture updated successfully!";
+        // Check for errors
+        if ($_FILES['eventBanner']['error'] == 0) {
+            // Delete the old photo if it exists
+            if ($currentPhoto && file_exists("../uploads/eventBanner/$currentPhoto")) {
+                unlink("../uploads/eventBanner/$currentPhoto");
+            }
+            
+            if (move_uploaded_file($tmp, '../uploads/eventBanner/' . $eventBanner)) {
+                $query = "UPDATE event SET event_banner = ? WHERE event_id = ?";
+                $stmt = $conn->prepare($query);
+                $stmt->execute([$eventBanner, $event_id]);
+
+                $_SESSION['success'] = "Profile picture updated successfully!";
+            } else {
+                $_SESSION['error'] = "Failed to upload photo.";
+            }
         } else {
             $_SESSION['error'] = "Failed to upload photo.";
         }
-    } else {
-        $_SESSION['error'] = "Failed to upload photo.";
-    }
 
-    header('Location: createEvent.php');
-    exit();
-}
+        header('Location: createEvent.php');
+        exit();
+    }
 ?>
